refactor(webapp): split NodeCard component
